+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - TT-1922 - activity-stream-integration
 - TT-1975 - implement rate-limiting
 - TT-2043 - upgrade notify client
+- TT-2062 - strip tags from email
 
 ### Fixed bugs:
 - No ticket - Upgrade django to 1.11.23 to fix vulnerability